requirements.txt
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0
plotly==5.19.0
streamlit==1.31.1
yfinance==0.2.36
